Draft: move more functions to draftgeoutils.arcs
